requirements.txt
cycler==0.10.0
Cython==0.29.23
dataclasses==0.6
evdev==1.4.0
future==0.18.2
imutils==0.5.3
kiwisolver==1.3.1
mss==6.1.0
opencv-python==4.5.1.48
pynput==1.7.3
pyparsing==2.4.7
python-dateutil==2.8.1
python-xlib==0.29
six==1.15.0
typing-extensions==3.7.4.3
#
torch==1.7.0
torchvision==0.8.0
# 
matplotlib==3.4.1
numpy==1.19.4
Pillow==8.2.0